refactor: Move `DatasetsSearchFilters` to `src.search.models`
